Move collision checking to CollisionManager class
