+++ pong.h
@@ -1,34 +1,22 @@
 #pragma once
 
+#include "collisionmanager.h"
 #include "gameobject.h"
 
 #include <vector>
 
-
 namespace pong
 {
-
-struct CollisionPair
-{
-    int mFirstGameObjectId;
-    int mSecondGameObjectId;
-};
-
 
 class Pong
 {
 private:
     std::vector<std::unique_ptr<GameObject>>mGameObjects {};
-    std::vector<CollisionPair>mCurrentCollisions {};
-
-    bool IsCurrentlyColliding(int firstGameObjectId, int secondGameObjectId) const;
-    bool IsCollisionPairValid(const CollisionPair& pair, int firstGameObjectId, int secondGameObjectId) const;
-    void RemoveGameObjectCollisionPair(int firstGameObjectId, int secondGameObjectId);
+    CollisionManager mCollisionManager {};
 
 public:
     void PongInit();
     void PongGameLoop();
-    void CheckForCollisions();
 };
 
 } // namespace pong
